+++ packages/cms-core/src/lib/components/admin/fields/ImageField.svelte
@@ -501,7 +501,7 @@
 										showOverride = false;
 									}}
 								>
-									Remove
+									Hide
 								</button>
 							{/if}
 						</div>